Refactor by adding schema into validation file
